Procfile

+++ resources/views/items/items.blade.php
@@ -11,20 +11,22 @@
                             @else
                                 <h2 class="item-title">{{ $item->name }}</h2>
                             @endif
-                
-                        <img border="0" src="/storage/image/{{$item->image_url}}" width="128" height="128" alt="">
+                                    <img border="0" src="/storage/image/{{$item->image_url}}" width="128" height="128" alt="">
+        
                         <h2><th>価格</th>
-                        {!! nl2br(e($item->price)) !!}</h2>
+                        {!! nl2br(e($item->price)) !!}円</h2>
                     </div>
                     <div>
-                            {!! Form::open(['route' => ['items.destroy', $item->id], 'method' => 'delete']) !!}
-                                {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-xs']) !!}
-                            {!! Form::close() !!}
-                            
-                            {!! Form::open(['route' => ['items.edit', $item->id], 'method' => 'get']) !!}
-                                {!! Form::submit('Edit', ['class' => 'btn btn-info btn-xs']) !!}
-                            {!! Form::close() !!}
-                </div>        
+                    @if (Auth::user()->id == $item->user_id)
+                        {!! Form::open(['route' => ['items.destroy', $item->id], 'method' => 'delete']) !!}
+                            {!! Form::submit('商品を削除する', ['class' => 'btn btn-danger btn-xs']) !!}
+                        {!! Form::close() !!}
+                        
+                        {!! Form::open(['route' => ['items.edit', $item->id], 'method' => 'get']) !!}
+                            {!! Form::submit('商品を編集する', ['class' => 'btn btn-info btn-xs']) !!}
+                        {!! Form::close() !!}
+                    @endif
+                </div>
             </div>
         </div>
     </li>
